requirements.txt
jupyter==1.0.0
numpy==1.18.1
pandas==1.0.1
PyAudio==0.2.11
python-telegram-bot==12.4.2
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
tenacity==6.1.0
torch==1.4.0
torchaudio==0.4.0
torchvision==0.5.0
tqdm==4.42.1
click==7.1.1
pytest==5.4.1
mock==4.0.2
